backend/requirements.txt
# ============================================
# Backend Requirements for Graph Agent
# ============================================

# Core AWS packages for Bedrock AgentCore
boto3>=1.34.0
boto3-stubs[bedrock-agent-runtime,bedrock-runtime]>=1.34.0

# Neo4j driver (already in root requirements but ensuring version)
neo4j>=5.15.0

# Environment and utilities
python-dotenv>=1.0.0
pydantic>=2.5.0

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Logging
structlog>=23.2.0